rework LookupPostcode to return []Address and extract inline JSON to testdata
