+++ .secrets.baseline
@@ -148,7 +148,7 @@
         "filename": "app/internal/place/client.go",
         "hashed_secret": "fca71afec681b7c2932610046e8e524820317e47",
         "is_verified": false,
-        "line_number": 76,
+        "line_number": 45,
         "is_secret": false
       }
     ],
@@ -243,5 +243,5 @@
       }
     ]
   },
-  "generated_at": "2022-10-13T15:07:03Z"
+  "generated_at": "2022-10-13T16:46:56Z"
 }
