+++ app/internal/page/choose_attorneys_address.go
@@ -49,17 +49,13 @@ func ChooseAttorneysAddress(logger Logger, tmpl template.Template, addressClient
 
 			if data.Form.Action == "lookup" && len(data.Errors) == 0 ||
 				data.Form.Action == "select" && len(data.Errors) > 0 {
-				response, err := addressClient.LookupPostcode(r.Context(), data.Form.LookupPostcode)
+				addresses, err := addressClient.LookupPostcode(r.Context(), data.Form.LookupPostcode)
 				if err != nil {
 					logger.Print(err)
 					data.Errors["lookup-postcode"] = "couldNotLookupPostcode"
 				}
 
-				if response.TotalResults > 0 {
-					data.Addresses = TransformAddressDetailsToAddresses(response.Results)
-				} else {
-					data.Addresses = []place.Address{}
-				}
+				data.Addresses = addresses
 			}
 		}
