+++ app/internal/page/data_test.go
@@ -6,8 +6,6 @@ import (
 	"testing"
 	"time"
 
-	"github.com/ministryofjustice/opg-modernising-lpa/internal/place"
-
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/mock"
 )
@@ -16,175 +14,6 @@ func TestReadDate(t *testing.T) {
 	date := readDate(time.Date(2020, time.March, 12, 0, 0, 0, 0, time.Local))
 
 	assert.Equal(t, Date{Day: "12", Month: "3", Year: "2020"}, date)
-}
-
-func TestTransformAddressDetailsToAddress(t *testing.T) {
-	testCases := []struct {
-		name string
-		ad   place.AddressDetails
-		want place.Address
-	}{
-		{
-			"Building number no building name",
-			place.AddressDetails{
-				Address:           "1, MELTON ROAD, BIRMINGHAM, B14 7ET",
-				BuildingName:      "",
-				BuildingNumber:    "1",
-				ThoroughFareName:  "MELTON ROAD",
-				DependentLocality: "",
-				Town:              "BIRMINGHAM",
-				Postcode:          "B14 7ET",
-			},
-			place.Address{Line1: "1 MELTON ROAD", Line2: "", Line3: "", TownOrCity: "BIRMINGHAM", Postcode: "B14 7ET"},
-		},
-		{
-			"Building name no building number",
-			place.AddressDetails{
-				Address:           "1A, MELTON ROAD, BIRMINGHAM, B14 7ET",
-				BuildingName:      "1A",
-				BuildingNumber:    "",
-				ThoroughFareName:  "MELTON ROAD",
-				DependentLocality: "",
-				Town:              "BIRMINGHAM",
-				Postcode:          "B14 7ET",
-			},
-			place.Address{Line1: "1A", Line2: "MELTON ROAD", Line3: "", TownOrCity: "BIRMINGHAM", Postcode: "B14 7ET"},
-		},
-		{
-			"Building name and building number",
-			place.AddressDetails{
-				Address:           "MELTON HOUSE, 2 MELTON ROAD, BIRMINGHAM, B14 7ET",
-				BuildingName:      "MELTON HOUSE",
-				BuildingNumber:    "2",
-				ThoroughFareName:  "MELTON ROAD",
-				DependentLocality: "",
-				Town:              "BIRMINGHAM",
-				Postcode:          "B14 7ET",
-			},
-			place.Address{Line1: "MELTON HOUSE", Line2: "2 MELTON ROAD", Line3: "", TownOrCity: "BIRMINGHAM", Postcode: "B14 7ET"},
-		},
-		{
-			"Dependent Locality building number",
-			place.AddressDetails{
-				Address:           "3, MELTON ROAD, BIRMINGHAM, B14 7ET",
-				BuildingName:      "",
-				BuildingNumber:    "3",
-				ThoroughFareName:  "MELTON ROAD",
-				DependentLocality: "KINGS HEATH",
-				Town:              "BIRMINGHAM",
-				Postcode:          "B14 7ET",
-			},
-			place.Address{Line1: "3 MELTON ROAD", Line2: "KINGS HEATH", Line3: "", TownOrCity: "BIRMINGHAM", Postcode: "B14 7ET"},
-		},
-		{
-			"Dependent Locality building name",
-			place.AddressDetails{
-				Address:           "MELTON HOUSE, MELTON ROAD, KINGS HEATH, BIRMINGHAM, B14 7ET",
-				BuildingName:      "MELTON HOUSE",
-				BuildingNumber:    "",
-				ThoroughFareName:  "MELTON ROAD",
-				DependentLocality: "KINGS HEATH",
-				Town:              "BIRMINGHAM",
-				Postcode:          "B14 7ET",
-			},
-			place.Address{Line1: "MELTON HOUSE", Line2: "MELTON ROAD", Line3: "KINGS HEATH", TownOrCity: "BIRMINGHAM", Postcode: "B14 7ET"},
-		},
-		{
-			"Dependent Locality building name and building number",
-			place.AddressDetails{
-				Address:           "MELTON HOUSE, 5 MELTON ROAD, KINGS HEATH BIRMINGHAM, B14 7ET",
-				BuildingName:      "MELTON HOUSE",
-				BuildingNumber:    "5",
-				ThoroughFareName:  "MELTON ROAD",
-				DependentLocality: "KINGS HEATH",
-				Town:              "BIRMINGHAM",
-				Postcode:          "B14 7ET",
-			},
-			place.Address{Line1: "MELTON HOUSE", Line2: "5 MELTON ROAD", Line3: "KINGS HEATH", TownOrCity: "BIRMINGHAM", Postcode: "B14 7ET"},
-		},
-	}
-
-	for _, tc := range testCases {
-		t.Run(tc.name, func(t *testing.T) {
-			assert.Equal(t, tc.want, TransformAddressDetailsToAddress(tc.ad))
-		})
-	}
-}
-
-func TestTransformAddressDetailsToAddresses(t *testing.T) {
-	ads := []place.AddressDetails{
-		{
-			Address:           "1, MELTON ROAD, BIRMINGHAM, B14 7ET",
-			BuildingName:      "",
-			BuildingNumber:    "1",
-			ThoroughFareName:  "MELTON ROAD",
-			DependentLocality: "",
-			Town:              "BIRMINGHAM",
-			Postcode:          "B14 7ET",
-		},
-		{
-			Address:           "2, MELTON ROAD, BIRMINGHAM, B14 7ET",
-			BuildingName:      "",
-			BuildingNumber:    "2",
-			ThoroughFareName:  "MELTON ROAD",
-			DependentLocality: "",
-			Town:              "BIRMINGHAM",
-			Postcode:          "B14 7ET",
-		},
-	}
-
-	want := []place.Address{
-		{Line1: "1 MELTON ROAD", Line2: "", Line3: "", TownOrCity: "BIRMINGHAM", Postcode: "B14 7ET"},
-		{Line1: "2 MELTON ROAD", Line2: "", Line3: "", TownOrCity: "BIRMINGHAM", Postcode: "B14 7ET"},
-	}
-
-	got := TransformAddressDetailsToAddresses(ads)
-
-	assert.Equal(t, want, got)
-}
-
-func TestAddress(t *testing.T) {
-	t.Run("String", func(t *testing.T) {
-		testCases := []struct {
-			name    string
-			address place.Address
-			want    string
-		}{
-			{
-				"All props set",
-				place.Address{
-					Line1:      "Line 1",
-					Line2:      "Line 2",
-					Line3:      "Line 3",
-					TownOrCity: "Town",
-					Postcode:   "Postcode",
-				},
-				"Line 1, Line 2, Line 3, Town, Postcode",
-			},
-			{
-				"Some props set",
-				place.Address{
-					Line1:      "Line 1",
-					Line2:      "",
-					Line3:      "Line 3",
-					TownOrCity: "Town",
-					Postcode:   "",
-				},
-				"Line 1, Line 3, Town",
-			},
-			{
-				"No props set",
-				place.Address{},
-				"",
-			},
-		}
-
-		for _, tc := range testCases {
-			t.Run(tc.name, func(t *testing.T) {
-				assert.Equal(t, tc.want, tc.address.String())
-			})
-		}
-	})
 }
 
 type mockDataStore struct {
