+++ app/internal/page/your_address.go
@@ -47,17 +47,13 @@ func YourAddress(logger Logger, tmpl template.Template, addressClient AddressCli
 
 			if data.Form.Action == "lookup" && len(data.Errors) == 0 ||
 				data.Form.Action == "select" && len(data.Errors) > 0 {
-				response, err := addressClient.LookupPostcode(r.Context(), data.Form.LookupPostcode)
+				addresses, err := addressClient.LookupPostcode(r.Context(), data.Form.LookupPostcode)
 				if err != nil {
 					logger.Print(err)
 					data.Errors["lookup-postcode"] = "couldNotLookupPostcode"
 				}
 
-				if response.TotalResults > 0 {
-					data.Addresses = TransformAddressDetailsToAddresses(response.Results)
-				} else {
-					data.Addresses = []place.Address{}
-				}
+				data.Addresses = addresses
 			}
 		}
