+++ app/internal/place/client_test.go
@@ -5,6 +5,7 @@ import (
 	"errors"
 	"net/http"
 	"net/http/httptest"
+	"os"
 	"testing"
 
 	"github.com/stretchr/testify/mock"
@@ -123,42 +124,36 @@ func (m *mockDoer) Do(req *http.Request) (*http.Response, error) {
 	return args.Get(0).(*http.Response), args.Error(1)
 }
 
-func TestFindAddress(t *testing.T) {
+func TestLookupPostcode(t *testing.T) {
+	multipleAddressJson, _ := os.ReadFile("testdata/postcode-multiple-addresses.json")
+	noAddressJson, _ := os.ReadFile("testdata/postcode-no-addresses.json")
+
 	ctx := context.Background()
 
 	testCases := []struct {
-		name                   string
-		postcode               string
-		queryPostcode          string
-		responseJson           string
-		expectedResponseObject PostcodeLookupResponse
+		name          string
+		postcode      string
+		queryPostcode string
+		responseJson  string
+		want          []Address
 	}{
 		{
 			"Multiple results",
 			"B14 7ET",
 			"B147ET",
-			multipleResultsJson,
-			PostcodeLookupResponse{
-				TotalResults: 2,
-				Results: []AddressDetails{
-					{
-						Address:           "123, MELTON ROAD, BIRMINGHAM, B14 7ET",
-						BuildingName:      "",
-						BuildingNumber:    "123",
-						ThoroughFareName:  "MELTON ROAD",
-						DependentLocality: "",
-						Town:              "BIRMINGHAM",
-						Postcode:          "B14 7ET",
-					},
-					{
-						Address:           "87A, MELTON ROAD, KINGS HEATH, BIRMINGHAM, B14 7ET",
-						BuildingName:      "87A",
-						BuildingNumber:    "",
-						ThoroughFareName:  "MELTON ROAD",
-						DependentLocality: "KINGS HEATH",
-						Town:              "BIRMINGHAM",
-						Postcode:          "B14 7ET",
-					},
+			string(multipleAddressJson),
+			[]Address{
+				{
+					Line1:      "123 MELTON ROAD",
+					TownOrCity: "BIRMINGHAM",
+					Postcode:   "B14 7ET",
+				},
+				{
+					Line1:      "87A",
+					Line2:      "MELTON ROAD",
+					Line3:      "KINGS HEATH",
+					TownOrCity: "BIRMINGHAM",
+					Postcode:   "B14 7ET",
 				},
 			},
 		},
@@ -166,8 +161,8 @@ func TestFindAddress(t *testing.T) {
 			"No results",
 			"  X XX XX X ",
 			"XXXXXX",
-			noResultsJson,
-			PostcodeLookupResponse{TotalResults: 0},
+			string(noAddressJson),
+			nil,
 		},
 	}
 
@@ -186,7 +181,7 @@ func TestFindAddress(t *testing.T) {
 			client := NewClient(server.URL, "fake-api-key", server.Client())
 			results, err := client.LookupPostcode(ctx, tc.postcode)
 
-			assert.Equal(t, tc.expectedResponseObject, results)
+			assert.Equal(t, tc.want, results)
 			assert.Nil(t, err)
 		})
 	}
@@ -232,45 +227,139 @@ func TestFindAddress(t *testing.T) {
 	})
 }
 
-func TestPostcodeLookupResponse(t *testing.T) {
-	t.Run("UnmarshalJSON", func(t *testing.T) {
-		t.Run("simplifies JSON structure in PostcodeLookupResponse object", func(t *testing.T) {
-			want := PostcodeLookupResponse{
-				TotalResults: 2,
-				Results: []AddressDetails{
-					{
-						Address:           "123, MELTON ROAD, BIRMINGHAM, B14 7ET",
-						BuildingName:      "",
-						BuildingNumber:    "123",
-						ThoroughFareName:  "MELTON ROAD",
-						DependentLocality: "",
-						Town:              "BIRMINGHAM",
-						Postcode:          "B14 7ET",
-					},
-					{
-						Address:           "87A, MELTON ROAD, KINGS HEATH, BIRMINGHAM, B14 7ET",
-						BuildingName:      "87A",
-						BuildingNumber:    "",
-						ThoroughFareName:  "MELTON ROAD",
-						DependentLocality: "KINGS HEATH",
-						Town:              "BIRMINGHAM",
-						Postcode:          "B14 7ET",
-					},
+func TestAddress(t *testing.T) {
+	t.Run("String", func(t *testing.T) {
+		testCases := []struct {
+			name    string
+			address Address
+			want    string
+		}{
+			{
+				"All props set",
+				Address{
+					Line1:      "Line 1",
+					Line2:      "Line 2",
+					Line3:      "Line 3",
+					TownOrCity: "Town",
+					Postcode:   "Postcode",
 				},
-			}
+				"Line 1, Line 2, Line 3, Town, Postcode",
+			},
+			{
+				"Some props set",
+				Address{
+					Line1:      "Line 1",
+					Line2:      "",
+					Line3:      "Line 3",
+					TownOrCity: "Town",
+					Postcode:   "",
+				},
+				"Line 1, Line 3, Town",
+			},
+			{
+				"No props set",
+				Address{},
+				"",
+			},
+		}
 
-			plr := PostcodeLookupResponse{}
-			plr.UnmarshalJSON([]byte(multipleResultsJson))
-
-			assert.Equal(t, want, plr)
-		})
-
-		t.Run("returns an error on unsuccessful an unmarshal", func(t *testing.T) {
-			plr := PostcodeLookupResponse{}
-			err := plr.UnmarshalJSON([]byte("not json"))
-
-			assert.ErrorContains(t, err, "invalid character")
-		})
+		for _, tc := range testCases {
+			t.Run(tc.name, func(t *testing.T) {
+				assert.Equal(t, tc.want, tc.address.String())
+			})
+		}
 	})
+}
 
+func TestTransformAddressDetailsToAddress(t *testing.T) {
+	testCases := []struct {
+		name string
+		ad   addressDetails
+		want Address
+	}{
+		{
+			"Building number no building name",
+			addressDetails{
+				Address:           "1, MELTON ROAD, BIRMINGHAM, B14 7ET",
+				BuildingName:      "",
+				BuildingNumber:    "1",
+				ThoroughFareName:  "MELTON ROAD",
+				DependentLocality: "",
+				Town:              "BIRMINGHAM",
+				Postcode:          "B14 7ET",
+			},
+			Address{Line1: "1 MELTON ROAD", Line2: "", Line3: "", TownOrCity: "BIRMINGHAM", Postcode: "B14 7ET"},
+		},
+		{
+			"Building name no building number",
+			addressDetails{
+				Address:           "1A, MELTON ROAD, BIRMINGHAM, B14 7ET",
+				BuildingName:      "1A",
+				BuildingNumber:    "",
+				ThoroughFareName:  "MELTON ROAD",
+				DependentLocality: "",
+				Town:              "BIRMINGHAM",
+				Postcode:          "B14 7ET",
+			},
+			Address{Line1: "1A", Line2: "MELTON ROAD", Line3: "", TownOrCity: "BIRMINGHAM", Postcode: "B14 7ET"},
+		},
+		{
+			"Building name and building number",
+			addressDetails{
+				Address:           "MELTON HOUSE, 2 MELTON ROAD, BIRMINGHAM, B14 7ET",
+				BuildingName:      "MELTON HOUSE",
+				BuildingNumber:    "2",
+				ThoroughFareName:  "MELTON ROAD",
+				DependentLocality: "",
+				Town:              "BIRMINGHAM",
+				Postcode:          "B14 7ET",
+			},
+			Address{Line1: "MELTON HOUSE", Line2: "2 MELTON ROAD", Line3: "", TownOrCity: "BIRMINGHAM", Postcode: "B14 7ET"},
+		},
+		{
+			"Dependent Locality building number",
+			addressDetails{
+				Address:           "3, MELTON ROAD, BIRMINGHAM, B14 7ET",
+				BuildingName:      "",
+				BuildingNumber:    "3",
+				ThoroughFareName:  "MELTON ROAD",
+				DependentLocality: "KINGS HEATH",
+				Town:              "BIRMINGHAM",
+				Postcode:          "B14 7ET",
+			},
+			Address{Line1: "3 MELTON ROAD", Line2: "KINGS HEATH", Line3: "", TownOrCity: "BIRMINGHAM", Postcode: "B14 7ET"},
+		},
+		{
+			"Dependent Locality building name",
+			addressDetails{
+				Address:           "MELTON HOUSE, MELTON ROAD, KINGS HEATH, BIRMINGHAM, B14 7ET",
+				BuildingName:      "MELTON HOUSE",
+				BuildingNumber:    "",
+				ThoroughFareName:  "MELTON ROAD",
+				DependentLocality: "KINGS HEATH",
+				Town:              "BIRMINGHAM",
+				Postcode:          "B14 7ET",
+			},
+			Address{Line1: "MELTON HOUSE", Line2: "MELTON ROAD", Line3: "KINGS HEATH", TownOrCity: "BIRMINGHAM", Postcode: "B14 7ET"},
+		},
+		{
+			"Dependent Locality building name and building number",
+			addressDetails{
+				Address:           "MELTON HOUSE, 5 MELTON ROAD, KINGS HEATH BIRMINGHAM, B14 7ET",
+				BuildingName:      "MELTON HOUSE",
+				BuildingNumber:    "5",
+				ThoroughFareName:  "MELTON ROAD",
+				DependentLocality: "KINGS HEATH",
+				Town:              "BIRMINGHAM",
+				Postcode:          "B14 7ET",
+			},
+			Address{Line1: "MELTON HOUSE", Line2: "5 MELTON ROAD", Line3: "KINGS HEATH", TownOrCity: "BIRMINGHAM", Postcode: "B14 7ET"},
+		},
+	}
+
+	for _, tc := range testCases {
+		t.Run(tc.name, func(t *testing.T) {
+			assert.Equal(t, tc.want, tc.ad.transformToAddress())
+		})
+	}
 }
