+++ app/internal/place/testdata/postcode-multiple-addresses.json
@@ -1,0 +1,82 @@
+{
+  "header": {
+    "uri": "https://api.os.uk/search/places/v1/postcode?postcode=B147ET",
+    "query": "postcode=B147ET",
+    "offset": 0,
+    "totalresults": 2,
+    "format": "JSON",
+    "dataset": "DPA",
+    "lr": "EN,CY",
+    "maxresults": 100,
+    "epoch": "96",
+    "output_srs": "EPSG:27700"
+  },
+  "results": [ {
+    "DPA": {
+      "UPRN": "100071390703",
+      "UDPRN": "432175",
+      "ADDRESS": "123, MELTON ROAD, BIRMINGHAM, B14 7ET",
+      "BUILDING_NUMBER": "123",
+      "THOROUGHFARE_NAME": "MELTON ROAD",
+      "POST_TOWN": "BIRMINGHAM",
+      "POSTCODE": "B14 7ET",
+      "RPC": "1",
+      "X_COORDINATE": 407783.0,
+      "Y_COORDINATE": 281505.0,
+      "STATUS": "APPROVED",
+      "LOGICAL_STATUS_CODE": "1",
+      "CLASSIFICATION_CODE": "RD04",
+      "CLASSIFICATION_CODE_DESCRIPTION": "Terraced",
+      "LOCAL_CUSTODIAN_CODE": 4605,
+      "LOCAL_CUSTODIAN_CODE_DESCRIPTION": "BIRMINGHAM",
+      "COUNTRY_CODE": "E",
+      "COUNTRY_CODE_DESCRIPTION": "This record is within England",
+      "POSTAL_ADDRESS_CODE": "D",
+      "POSTAL_ADDRESS_CODE_DESCRIPTION": "A record which is linked to PAF",
+      "BLPU_STATE_CODE": "2",
+      "BLPU_STATE_CODE_DESCRIPTION": "In use",
+      "TOPOGRAPHY_LAYER_TOID": "osgb1000020369531",
+      "LAST_UPDATE_DATE": "10/02/2016",
+      "ENTRY_DATE": "16/04/2001",
+      "BLPU_STATE_DATE": "29/04/2013",
+      "LANGUAGE": "EN",
+      "MATCH": 1.0,
+      "MATCH_DESCRIPTION": "EXACT",
+      "DELIVERY_POINT_SUFFIX": "1Q"
+    }
+  }, {
+    "DPA": {
+      "UPRN": "100070449924",
+      "UDPRN": "432202",
+      "ADDRESS": "87A, MELTON ROAD, KINGS HEATH, BIRMINGHAM, B14 7ET",
+      "BUILDING_NAME": "87A",
+      "THOROUGHFARE_NAME": "MELTON ROAD",
+      "DEPENDENT_LOCALITY": "KINGS HEATH",
+      "POST_TOWN": "BIRMINGHAM",
+      "POSTCODE": "B14 7ET",
+      "RPC": "1",
+      "X_COORDINATE": 407799.0,
+      "Y_COORDINATE": 281591.0,
+      "STATUS": "APPROVED",
+      "LOGICAL_STATUS_CODE": "1",
+      "CLASSIFICATION_CODE": "RD04",
+      "CLASSIFICATION_CODE_DESCRIPTION": "Terraced",
+      "LOCAL_CUSTODIAN_CODE": 4605,
+      "LOCAL_CUSTODIAN_CODE_DESCRIPTION": "BIRMINGHAM",
+      "COUNTRY_CODE": "E",
+      "COUNTRY_CODE_DESCRIPTION": "This record is within England",
+      "POSTAL_ADDRESS_CODE": "D",
+      "POSTAL_ADDRESS_CODE_DESCRIPTION": "A record which is linked to PAF",
+      "BLPU_STATE_CODE": "2",
+      "BLPU_STATE_CODE_DESCRIPTION": "In use",
+      "TOPOGRAPHY_LAYER_TOID": "osgb1000020369554",
+      "LAST_UPDATE_DATE": "10/02/2016",
+      "ENTRY_DATE": "16/04/2001",
+      "BLPU_STATE_DATE": "29/04/2013",
+      "LANGUAGE": "EN",
+      "MATCH": 1.0,
+      "MATCH_DESCRIPTION": "EXACT",
+      "DELIVERY_POINT_SUFFIX": "3E"
+    }
+  }]
+}
