+++ app/internal/place/testdata/postcode-no-addresses.json
@@ -1,0 +1,14 @@
+{
+  "header": {
+    "uri": "https://api.os.uk/search/places/v1/postcode?postcode=XXXXXX",
+    "query": "postcode=XXXXXX",
+    "offset": 0,
+    "totalresults": 0,
+    "format": "JSON",
+    "dataset": "DPA",
+    "lr": "EN,CY",
+    "maxresults": 100,
+    "epoch": "96",
+    "output_srs": "EPSG:27700"
+  }
+}
